+++ src/scene-setup/init.js
@@ -1,0 +1,22 @@
+import { World } from './_World.js';
+import addDefaultObject from './_addDefaultObject.js';
+import { control_as_FPS, control_as_Orbit } from './_defaultCameraControl.js';
+
+
+const world = new World(document.getElementById('bg'));
+export default world;
+
+let stateToUpdate;
+stateToUpdate = addDefaultObject(
+    world.scene,
+);
+world.addSceneState(stateToUpdate);
+
+const isMobileDevice = /Mobi/i.test(window.navigator.userAgent);
+if(isMobileDevice){
+    control_as_Orbit(world.camera, world.renderer.domElement);
+}
+else {
+    stateToUpdate = control_as_FPS(world.camera, world.renderer.domElement);
+    world.addSceneState(stateToUpdate);
+}
